Added module name sanitizing to github downloading and build
