+++ src/jablkomods/downloading.go
@@ -11,8 +11,8 @@ package jablkomods
 import (
 	"net/http"
 	"io"
-	"archive/zip"
-	"path/filepath"
+	"bufio"
+	"io/ioutil"
 	"os"
 	"fmt"
 	"strings"
@@ -66,67 +66,35 @@ func downloadGithub(sourcePath string) error {
 	}
 
 	splitSource := strings.Split(sourcePath, "/")
+	installPath := GithubSourceToInstallDir(sourcePath)
 
 	_, err = Unzip("./tmp/" + sourcePath +  "/source.zip", "./" + splitSource[0] + "/" + splitSource[1])
 	if err != nil {
 		return err
 	}
+
+	// Replace module line in go.mod to include version
+	goModFile, err := os.Open(installPath + "/go.mod")
+	if err != nil {
+		return err
+	}
+
+	modScanner := bufio.NewScanner(goModFile)
+	modScanner.Split(bufio.ScanLines)
+	var modLines []string
+
+	for modScanner.Scan() {
+		tempLine := modScanner.Text()
+		if strings.HasPrefix(tempLine, "module") {
+			modLines = append(modLines, "module " + sourcePath)
+		} else {
+			modLines = append(modLines, tempLine)
+		}
+	}
+
+	goModFile.Close()
+
+	err = ioutil.WriteFile(sourcePath + "/go.mod", []byte(strings.Join(modLines, "\n")),0666)
+
 	return err
-}
-
-func Unzip(src string, dest string) ([]string, error) {
-	// From golangcode.com
-
-    var filenames []string
-
-    r, err := zip.OpenReader(src)
-    if err != nil {
-        return filenames, err
-    }
-    defer r.Close()
-
-    for _, f := range r.File {
-
-        // Store filename/path for returning and using later on
-        fpath := filepath.Join(dest, f.Name)
-
-        // Check for ZipSlip. More Info: http://bit.ly/2MsjAWE
-        if !strings.HasPrefix(fpath, filepath.Clean(dest)+string(os.PathSeparator)) {
-            return filenames, fmt.Errorf("%s: illegal file path", fpath)
-        }
-
-        filenames = append(filenames, fpath)
-
-        if f.FileInfo().IsDir() {
-            // Make Folder
-            os.MkdirAll(fpath, os.ModePerm)
-            continue
-        }
-
-        // Make File
-        if err = os.MkdirAll(filepath.Dir(fpath), os.ModePerm); err != nil {
-            return filenames, err
-        }
-
-        outFile, err := os.OpenFile(fpath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, f.Mode())
-        if err != nil {
-            return filenames, err
-        }
-
-        rc, err := f.Open()
-        if err != nil {
-            return filenames, err
-        }
-
-        _, err = io.Copy(outFile, rc)
-
-        // Close the file without defer to close before next iteration of loop
-        outFile.Close()
-        rc.Close()
-
-        if err != nil {
-            return filenames, err
-        }
-    }
-    return filenames, nil
 }
