+++ src/jablkomods/utils.go
@@ -7,6 +7,11 @@ package jablkomods
 
 import (
 	"strings"
+	"archive/zip"
+	"path/filepath"
+	"os"
+	"io"
+	"fmt"
 )
 
 func GithubSourceToURL(sourceStr string) string {
@@ -20,11 +25,62 @@ func GithubSourceToURL(sourceStr string) string {
 }
 
 func GithubSourceToInstallDir(sourceStr string) string {
-	splitSource := strings.Split(sourceStr, "/")
+	return sourceStr
+}
 
-	if strings.HasSuffix(sourceStr, "master") {
-		return splitSource[0] + "/" + splitSource[1] + "/" + splitSource[2] + "-" + splitSource[3]
-	} else {
-		return splitSource[0] + "/" + splitSource[1] + "/" + splitSource[2] + "-" + splitSource[3][1:len(splitSource[3])]
-	}
+func Unzip(src string, dest string) ([]string, error) {
+	// From golangcode.com
+
+    var filenames []string
+
+    r, err := zip.OpenReader(src)
+    if err != nil {
+        return filenames, err
+    }
+    defer r.Close()
+
+    for _, f := range r.File {
+
+        // Store filename/path for returning and using later on
+        fpath := filepath.Join(dest, f.Name)
+
+        // Check for ZipSlip. More Info: http://bit.ly/2MsjAWE
+        if !strings.HasPrefix(fpath, filepath.Clean(dest)+string(os.PathSeparator)) {
+            return filenames, fmt.Errorf("%s: illegal file path", fpath)
+        }
+
+        filenames = append(filenames, fpath)
+
+        if f.FileInfo().IsDir() {
+            // Make Folder
+            os.MkdirAll(fpath, os.ModePerm)
+            continue
+        }
+
+        // Make File
+        if err = os.MkdirAll(filepath.Dir(fpath), os.ModePerm); err != nil {
+            return filenames, err
+        }
+
+        outFile, err := os.OpenFile(fpath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, f.Mode())
+        if err != nil {
+            return filenames, err
+        }
+
+        rc, err := f.Open()
+        if err != nil {
+            return filenames, err
+        }
+
+        _, err = io.Copy(outFile, rc)
+
+        // Close the file without defer to close before next iteration of loop
+        outFile.Close()
+        rc.Close()
+
+        if err != nil {
+            return filenames, err
+        }
+    }
+    return filenames, nil
 }
